components/requirements.txt
altair==4.2.2
streamlit==1.43.2
google-generativeai==0.8.4